+++ apps/minifront/src/components/tx-details/index.tsx
@@ -28,19 +28,21 @@ export const TxDetails = () => {
 
   return (
     <FadeTransition className='flex min-h-[calc(100vh-122px)] flex-col items-stretch justify-start'>
-      <RestrictMaxWidth>
-        <div className='relative grid grid-std-spacing lg:grid-cols-3'>
-          <Card gradient className='flex-1 p-5 md:p-4 lg:col-span-2 lg:row-span-2 xl:p-5'>
-            <TxViewer txInfo={txInfo.data} />
-          </Card>
-          <EduInfoCard
-            className='row-span-1'
-            src='./incognito.svg'
-            label='Shielded Transactions'
-            content={EduPanel.SHIELDED_TRANSACTION}
-          />
-        </div>
-      </RestrictMaxWidth>
+      <div className='flex flex-col'>
+        <RestrictMaxWidth>
+          <div className='relative grid grid-std-spacing lg:grid-cols-3'>
+            <Card gradient className='flex-1 p-5 md:p-4 lg:col-span-2 lg:row-span-2 xl:p-5'>
+              <TxViewer txInfo={txInfo.data} />
+            </Card>
+            <EduInfoCard
+              className='row-span-1'
+              src='./incognito.svg'
+              label='Shielded Transactions'
+              content={EduPanel.SHIELDED_TRANSACTION}
+            />
+          </div>
+        </RestrictMaxWidth>
+      </div>
     </FadeTransition>
   );
 };