+++ packages/ui/src/AddressViewComponent/index.test.tsx
@@ -1,49 +1,14 @@
-import {
-  Address,
-  AddressIndex,
-  AddressView,
-  AddressView_Decoded,
-} from '@penumbra-zone/protobuf/penumbra/core/keys/v1/keys_pb';
 import { AddressViewComponent } from '.';
 import { describe, expect, it } from 'vitest';
 import { render } from '@testing-library/react';
 import { PenumbraUIProvider } from '../PenumbraUIProvider';
-
-const addressViewWithOneTimeAddress = new AddressView({
-  addressView: {
-    case: 'decoded',
-
-    value: new AddressView_Decoded({
-      address: new Address({ inner: new Uint8Array(80) }),
-      index: new AddressIndex({
-        account: 0,
-        // A one-time address is defined by a randomizer with at least one
-        // non-zero byte.
-        randomizer: new Uint8Array([1, 2, 3]),
-      }),
-    }),
-  },
-});
-
-const addressViewWithNormalAddress = new AddressView({
-  addressView: {
-    case: 'decoded',
-
-    value: new AddressView_Decoded({
-      address: new Address({ inner: new Uint8Array(80) }),
-      index: new AddressIndex({
-        account: 0,
-        randomizer: new Uint8Array([0, 0, 0]),
-      }),
-    }),
-  },
-});
+import { ADDRESS_VIEW_DECODED_ONE_TIME, ADDRESS_VIEW_DECODED } from '../utils/bufs';
 
 describe('<AddressViewComponent />', () => {
   describe('when `copyable` is `true`', () => {
     it('does not show the copy icon when the address is a one-time address', () => {
       const { queryByLabelText } = render(
-        <AddressViewComponent addressView={addressViewWithOneTimeAddress} copyable />,
+        <AddressViewComponent addressView={ADDRESS_VIEW_DECODED_ONE_TIME} copyable />,
         { wrapper: PenumbraUIProvider },
       );
 
@@ -52,7 +17,7 @@ describe('<AddressViewComponent />', () => {
 
     it('shows the copy icon when the address is not a one-time address', () => {
       const { queryByLabelText } = render(
-        <AddressViewComponent addressView={addressViewWithNormalAddress} copyable />,
+        <AddressViewComponent addressView={ADDRESS_VIEW_DECODED} copyable />,
         { wrapper: PenumbraUIProvider },
       );
 
@@ -63,7 +28,7 @@ describe('<AddressViewComponent />', () => {
   describe('when `copyable` is `false`', () => {
     it('does not show the copy icon when the address is a one-time address', () => {
       const { queryByLabelText } = render(
-        <AddressViewComponent addressView={addressViewWithOneTimeAddress} copyable={false} />,
+        <AddressViewComponent addressView={ADDRESS_VIEW_DECODED_ONE_TIME} copyable={false} />,
         { wrapper: PenumbraUIProvider },
       );
 
@@ -72,7 +37,7 @@ describe('<AddressViewComponent />', () => {
 
     it('does not show the copy icon when the address is not a one-time address', () => {
       const { queryByLabelText } = render(
-        <AddressViewComponent addressView={addressViewWithNormalAddress} copyable={false} />,
+        <AddressViewComponent addressView={ADDRESS_VIEW_DECODED} copyable={false} />,
         { wrapper: PenumbraUIProvider },
       );