+++ packages/query/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @penumbra-zone/query
 
+## 28.0.0
+
+### Patch Changes
+
+- Updated dependencies [598d148]
+  - @penumbra-zone/wasm@26.1.0
+
+## 27.0.0
+
+### Minor Changes
+
+- 94d2c8d: special case local genesis sync
+
+### Patch Changes
+
+- Updated dependencies [f8730e9]
+  - @penumbra-zone/getters@16.0.0
+  - @penumbra-zone/types@21.0.0
+  - @penumbra-zone/crypto-web@22.0.0
+  - @penumbra-zone/wasm@26.0.0
+
 ## 26.0.0
 
 ### Patch Changes