+++ packages/services/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @penumbra-zone/router
 
+## 31.0.0
+
+### Patch Changes
+
+- Updated dependencies [598d148]
+- Updated dependencies [e952e03]
+  - @penumbra-zone/wasm@26.1.0
+  - @penumbra-zone/storage@27.0.0
+  - @penumbra-zone/query@28.0.0
+
+## 30.0.0
+
+### Patch Changes
+
+- Updated dependencies [f8730e9]
+- Updated dependencies [94d2c8d]
+  - @penumbra-zone/getters@16.0.0
+  - @penumbra-zone/query@27.0.0
+  - @penumbra-zone/storage@26.0.0
+  - @penumbra-zone/types@21.0.0
+  - @penumbra-zone/crypto-web@22.0.0
+  - @penumbra-zone/wasm@26.0.0
+
 ## 29.0.0
 
 ### Patch Changes