+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @penumbra-zone/storage
 
+## 27.0.0
+
+### Patch Changes
+
+- e952e03: remove BSR dependencies, correctly specify @bufbuild and @connectrpc dependencies
+- Updated dependencies [598d148]
+  - @penumbra-zone/wasm@26.1.0
+
+## 26.0.0
+
+### Patch Changes
+
+- Updated dependencies [f8730e9]
+  - @penumbra-zone/getters@16.0.0
+  - @penumbra-zone/types@21.0.0
+  - @penumbra-zone/wasm@26.0.0
+
 ## 25.0.0
 
 ### Patch Changes